refactor(docker_conf): rename and remove unnecessary options
